+++ rebar/src/main/kotlin/io/github/pylonmc/rebar/guide/button/BackButton.kt
@@ -2,10 +2,10 @@ package io.github.pylonmc.rebar.guide.button
 
 import io.github.pylonmc.rebar.config.RebarConfig
 import io.github.pylonmc.rebar.content.guide.RebarGuide
+import io.github.pylonmc.rebar.content.guide.RebarGuide.Companion.guideHints
 import io.github.pylonmc.rebar.guide.pages.base.PagedGuidePage
 import io.github.pylonmc.rebar.item.builder.ItemStackBuilder
 import io.github.pylonmc.rebar.content.guide.RebarGuide.Companion.playGuideSound
-import io.github.pylonmc.rebar.guide.pages.settings.MainSettingsPage.guideHints
 import io.github.pylonmc.rebar.util.rebarKey
 import io.papermc.paper.datacomponent.DataComponentTypes
 import net.kyori.adventure.text.Component
