+++ rebar/src/main/kotlin/io/github/pylonmc/rebar/guide/button/ResearchButton.kt
@@ -1,12 +1,12 @@
 package io.github.pylonmc.rebar.guide.button
 
 import io.github.pylonmc.rebar.config.RebarConfig
+import io.github.pylonmc.rebar.content.guide.RebarGuide.Companion.guideHints
 import io.github.pylonmc.rebar.guide.pages.research.ResearchItemsPage
 import io.github.pylonmc.rebar.i18n.RebarArgument
 import io.github.pylonmc.rebar.item.builder.ItemStackBuilder
 import io.github.pylonmc.rebar.item.research.Research
 import io.github.pylonmc.rebar.content.guide.RebarGuide.Companion.playGuideSound
-import io.github.pylonmc.rebar.guide.pages.settings.MainSettingsPage.guideHints
 import io.github.pylonmc.rebar.item.research.Research.Companion.researchPoints
 import io.github.pylonmc.rebar.util.gui.unit.UnitFormat
 import io.github.pylonmc.rebar.util.rebarKey
